Enum raster::editor::ResizeMode

pub enum ResizeMode {
    Exact,
    ExactWidth,
    ExactHeight,
    Fit,
    Fill,
}

An enum for the various modes that can be used for resizing.

Variants
Exact

Resize image to exact dimensions ignoring aspect ratio.

ExactWidth

Resize image to exact width. Height parameter is ignored and is auto calculated instead.

ExactHeight

Resize image to exact height. Width parameter is ignored and is auto calculated instead.

Fit

Resize an image to fit within the given width and height.

Fill

Resize image to fill all the space in the given dimension. Excess parts are cropped.